Fix bug of afb_data_array_convert
The use of the trailer in afb_data_array_convert can lead to
crash because the trailer isn't a valid afb_type but a weird
pointer and argc is not always checked.

Using afb_req_param_convert also removes the need of releasing
conversion, things that wasn't always cleanly done.